lastgenre: Move tree helpers to where they are used
